[storage service]: public bucket. [general]: get url and create video metadata

+++ video-service/src/video/entities/video.entity.ts
@@ -51,7 +51,7 @@ export class Video {
   @Column({ nullable: true })
   thumbnailUrl?: string;
 
-  @Column({ type: 'bigint', nullable: true })
+  @Column({ type: 'float', nullable: true })
   duration?: number;
 
   @Column({ default: false })
